requirements.txt
attrs==23.1.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.7
Faker==16.1.0
inflection==0.5.1
jsonschema==4.17.3
mysqlclient==2.2.0
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
uritemplate==4.1.1
wincertstore==0.2
